requirements.txt
fastapi==0.110.2
uvicorn==0.30.0
transformers==4.46.2
torch==2.6.0
langid==1.1.6
python-docx==1.1.2
PyPDF2==3.0.1
sentencepiece==0.2.0
sacremoses==0.1.1
protobuf==4.25.3
gradio==4.44.0
requests==2.32.3
